+++ resources/js/Components/Galleries/GalleryPage/GallerySelectTemplate/GallerySelectTemplate.test.tsx
@@ -1,5 +1,5 @@
 import { GallerySelectTemplate } from "./GallerySelectTemplate";
-import * as useDarkModeContext from "@/Contexts/DarkModeContex";
+import * as useDarkModeContext from "@/Contexts/DarkModeContext";
 import { render, screen } from "@/Tests/testing-library";
 
 describe("GallerySelectTemplate", () => {
