+++ resources/js/Components/Layout/AuthOverlay/AuthOverlay.tsx
@@ -12,7 +12,7 @@ import { Heading } from "@/Components/Heading";
 import { Overlay } from "@/Components/Layout/Overlay/Overlay";
 import { Toast } from "@/Components/Toast";
 import { useAuth } from "@/Contexts/AuthContext";
-import { useDarkModeContext } from "@/Contexts/DarkModeContex";
+import { useDarkModeContext } from "@/Contexts/DarkModeContext";
 import { useMetaMaskContext } from "@/Contexts/MetaMaskContext";
 import { AuthConnectWallet, AuthConnectWalletDark, AuthInstallWallet } from "@/images";
 import { isTruthy } from "@/Utils/is-truthy";
